refactor: reduce email_view.dart file size
