+++ lib/utils/html_has_images.dart
@@ -1,0 +1,3 @@
+bool htmlHasImages(String html) {
+  return RegExp(r'<img[\s>/]', caseSensitive: false).hasMatch(html);
+}
