+++ lib/views/email/widgets/html_body_view.dart
@@ -1,0 +1,75 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
+import 'package:nostr_mail_client/utils/confirm_open_link.dart';
+import 'package:nostr_mail_client/utils/html_has_images.dart';
+import 'package:nostr_mail_client/views/email/email_controller.dart';
+
+class HtmlBodyView extends StatelessWidget {
+  final String htmlBody;
+
+  const HtmlBodyView({super.key, required this.htmlBody});
+
+  @override
+  Widget build(BuildContext context) {
+    final hasImages = htmlHasImages(htmlBody);
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        if (hasImages && !EmailController.to.showImages)
+          Container(
+            width: double.infinity,
+            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+            margin: const EdgeInsets.only(bottom: 16),
+            decoration: BoxDecoration(
+              color: Theme.of(context).colorScheme.surfaceContainerHighest,
+              borderRadius: BorderRadius.circular(8),
+            ),
+            child: Row(
+              children: [
+                Icon(
+                  Icons.image_not_supported_outlined,
+                  size: 18,
+                  color: Theme.of(context).colorScheme.onSurfaceVariant,
+                ),
+                const SizedBox(width: 8),
+                Expanded(
+                  child: Text(
+                    'Images are hidden for privacy',
+                    style: TextStyle(
+                      fontSize: 13,
+                      color: Theme.of(context).colorScheme.onSurfaceVariant,
+                    ),
+                  ),
+                ),
+                TextButton(
+                  onPressed: () {
+                    EmailController.to.showImages = true;
+                    EmailController.to.update();
+                  },
+                  child: const Text('Load images'),
+                ),
+              ],
+            ),
+          ),
+        SelectionArea(
+          child: HtmlWidget(
+            htmlBody,
+            key: ValueKey(EmailController.to.showImages),
+            customWidgetBuilder: EmailController.to.showImages
+                ? null
+                : (element) {
+                    if (element.localName == 'img') {
+                      return const SizedBox.shrink();
+                    }
+                    return null;
+                  },
+            onTapUrl: (url) {
+              confirmOpenLink(url);
+              return true;
+            },
+          ),
+        ),
+      ],
+    );
+  }
+}
