+++ lib/views/email/widgets/recipient_bridge_badge_view.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:nostr_mail_client/views/email/email_controller.dart';
+import 'package:nostr_mail_client/widgets/nostr_avatar.dart';
+
+class RecipientBridgeBadgeView extends StatelessWidget {
+  const RecipientBridgeBadgeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return NostrAvatar(
+      pubkey: EmailController.to.email!.recipientPubkey,
+      metadata: EmailController.to.recipientBridgeMetadata,
+      radius: 7,
+    );
+  }
+}
